+++ packages/pebble/src/compiler/AstCompiler/internal/exprs/_compileVarAccessExpr.ts
@@ -18,7 +18,6 @@ export function _compileVarAccessExpr(
             expr.range
         );
 
-        console.trace( ctx.scope.allVariables(), expr.text );
         return ctx.error(
             DiagnosticCode._0_is_not_defined,
             expr.range, expr.text